Backend_prediction/requirements.txt
pathlib == 1.0.1
uvicorn == 0.21.1
fastapi == 0.95.0
fastai == 2.7.12
numpy == 1.24.2
Pillow == 9.4.0
opencv-contrib-python-headless
boto3 == 1.26.101
python-multipart == 0.0.6